% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bivariate_choropleth.R
\name{addBivariateChoropleth}
\alias{addBivariateChoropleth}
\title{Add a Bivariate Choropleth to a Leaflet Map}
\usage{
addBivariateChoropleth(
  map,
  map_data,
  var1_name,
  var2_name,
  ntiles = 3,
  var1_label = NA,
  var2_label = NA,
  label_arrows = TRUE,
  region_names = NA,
  paletteFunction = pals::stevens.pinkblue,
  ...
)
}
\arguments{
\item{map}{A Leaflet map created by `leaflet::leaflet()`.}

\item{...}{}
}
\value{

}
\description{
Add a Bivariate Choropleth to a Leaflet Map
}
