% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bivariate_choropleth.R
\name{bivariate_choropleth}
\alias{bivariate_choropleth}
\title{Make a Bivariate Choropleth with ggplot2, cowplot, and ggiraph}
\usage{
bivariate_choropleth(
  df,
  var1_name,
  var2_name,
  var1_label = "var1",
  var2_label = "var2",
  polygon_names,
  na_values = c("zero", "drop_na"),
  plot_title = NULL,
  plot_subtitle = NULL,
  legend_xpos = 0.75,
  legend_ypos = 0.1,
  legend_size = 0.15,
  interactive = TRUE,
  label_arrows = TRUE
)
}
\arguments{
\item{label_arrows}{}
}
\value{

}
\description{
Make a Bivariate Choropleth with ggplot2, cowplot, and ggiraph
}
