% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bivariate_choropleth.R
\name{addBivariateChoropleth}
\alias{addBivariateChoropleth}
\title{Add a Bivariate Choropleth to a Leaflet Map}
\usage{
addBivariateChoropleth(
  map,
  map_data,
  var1_name,
  var2_name,
  ntiles = 3,
  var1_label = NA,
  var2_label = NA,
  label_arrows = TRUE,
  region_names = NA,
  add_legend = TRUE,
  paletteFunction = pals::stevens.pinkblue,
  ...
)
}
\arguments{
\item{map}{A Leaflet map created by `leaflet::leaflet()`.}

\item{map_data}{`sf` object with geometry features and two columns of data to plot with a bivariate choropleth.}

\item{var1_name}{Column name of the first variable.}

\item{var2_name}{Column name of the second variable.}

\item{ntiles}{Number of ntiles. 3 works well.}

\item{var1_label}{Human-readable label for variable 1.}

\item{var2_label}{Human-readable label for variable 2.}

\item{label_arrows}{Boolean, defaults TRUE. Should we show nice arrows on the legend?}

\item{region_names}{Column with region names.}

\item{add_legend}{Boolean, defaults TRUE. Should we add a legend?}

\item{paletteFunction}{Bivariate palette function. Defaults to `pals::stevens.pinkblue`.}

\item{...}{Other arguments passed to Leaflet.}
}
\value{
A Leaflet object with a bivariate choropleth.
}
\description{
Please note this function is a work in progress and this documentation needs
improving.
}
\details{
Please submit any bugs or issues on GitHub:
https://github.com/chris31415926535/bivariatechoropleths
}
\examples{
\dontrun{
leaflet::leaflet() \%>\%
  leaflet::addTiles() \%>\%
  bivariatechoropleths::addBivariateChoropleth(
    map_data = bivariatechoropleths::renfrew_county,
    var1_name = pop_2016,
    var2_name = median_household_income_2015,
    ntiles= 3,
    var1_label = "Population, 2016",
    var2_label = "Median Household\nIncome, 2015",
    region_name = "CSDNAME",
    weight = 1,
    fillOpacity = 0.7,
    color = "grey",
    highlightOptions = leaflet::highlightOptions(color = "orange",
                                                 weight = 2,
                                                 opacity = 1))
#'     }
}
